server/design-rules-engine/main.py
import os
from fastapi import FastAPI, Request, Response, HTTPException
from dotenv import load_dotenv
# Import Pydantic models later for request/response validation
# from pydantic import BaseModel
# Import rule-checking libraries later

# Load environment variables from .env file
load_dotenv()

app = FastAPI(
    title="Design Rules Engine",
    description="Validates generated website content/styles against design rules (consistency, contrast, etc.).",
    version="1.0.0"
)

# --- Data Models (Example - Define based on actual needs) ---
# class DesignInput(BaseModel):
#     template_id: str
#     generated_content: dict # Structure representing pages, sections, styles
#     branding: dict | None = None

# class ValidationResult(BaseModel):
#     passed: bool
#     issues: list[dict] = [] # List of identified design rule violations

# --- API Endpoints ---

@app.get("/")
async def read_root():
    """ Basic health check endpoint. """
    return {"message": "Design Rules Engine is running!"}

# Example endpoint (replace with actual validation logic)
# @app.post("/validate-design", response_model=ValidationResult)
# async def validate_design(data: DesignInput):
#     """
#     Placeholder for design rule validation logic.
#     In a real implementation, this would check consistency, contrast, hierarchy, etc.
#     """
#     print(f"Received request to validate design for template: {data.template_id}")
#     # TODO: Implement actual design rule checks.

#     # Placeholder response (assuming it passes for now)
#     return ValidationResult(passed=True, issues=[])


# --- Run the server (for local development) ---
if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", "3010")) # Default to port 3010 for design rules engine
    print(f"Starting Design Rules Engine on http://localhost:{port}")
    # Use reload=True for development
    uvicorn.run("main:app", host="0.0.0.0", port=port, reload=True)
