server/design-rules-engine/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
# Add specific rule/validation libraries later if needed
